requirements.txt
langchain-community>=0.2.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.7.4
PyYAML>=6.0.1
pypdf>=4.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
PyMuPDF>=1.24.9
fastapi>=0.115.0
uvicorn>=0.30.0
